git-svn-id: http://svn.code.sf.net/p/qlandkartegt/code/QLandkarteGT/trunk@2892 d0c010d8-bb82-477e-a075-112140230451

+++ src/CMapQMAPExport.cpp
@@ -377,7 +377,7 @@ void CMapQMAPExport::startQLM()
                 {
                     job_t tmpJob = job;
 
-                    div = (quint64)sqrt(div/(MAXENDURA));
+                    div = (quint64)sqrt((float)div/(MAXENDURA));
                     div += 1;
 
                     int tilesX = job.width/div;
